+++ src/main/java/uk/gov/ons/ctp/response/sample/endpoint/SampleSummaryDistributionEndpoint.java
@@ -2,7 +2,6 @@ package uk.gov.ons.ctp.response.sample.endpoint;
 
 import static net.logstash.logback.argument.StructuredArguments.kv;
 
-import java.util.List;
 import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,8 +11,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.ons.ctp.response.sample.domain.model.SampleUnit;
-import uk.gov.ons.ctp.response.sample.service.SampleService;
+import uk.gov.ons.ctp.response.sample.service.NoSampleUnitsInSampleSummaryException;
+import uk.gov.ons.ctp.response.sample.service.SampleSummaryDistributionService;
+import uk.gov.ons.ctp.response.sample.service.UnknownSampleSummaryException;
 
 @RestController
 @RequestMapping(value = "/distribute", produces = "application/json")
@@ -21,44 +21,26 @@ public class SampleSummaryDistributionEndpoint {
 
   private static final Logger LOG =
       LoggerFactory.getLogger(SampleSummaryDistributionEndpoint.class);
-  @Autowired private SampleService sampleService;
+
+  @Autowired private SampleSummaryDistributionService sampleSummaryDistributionService;
 
   /**
    * Distributes all the SampleUnits for a SampleSummary to case. Distributing here means to tell
    * case about all the sampleUnits so cases can be made against each one to track survey completion
    * progress.
    *
-   * @param collectionExerciseId The collection exercise id for the sample summary
    * @param sampleSummaryId The sample summary ID used to find all the sample units to distribute
    * @return A HTTP response
    */
-  @RequestMapping(
-      value = "/collection-exercise/{collectionExerciseId}/samplesummary/{sampleSummaryId}",
-      method = RequestMethod.GET)
-  public ResponseEntity<Void> distribute(
-      @PathVariable("collectionExerciseId") String collectionExerciseId,
-      @PathVariable("sampleSummaryId") UUID sampleSummaryId) {
+  @RequestMapping(value = "/{sampleSummaryId}", method = RequestMethod.GET)
+  public ResponseEntity<Void> distribute(@PathVariable("sampleSummaryId") UUID sampleSummaryId) {
 
-    LOG.info("Distributing sample units to case", kv("collectionExerciseId", collectionExerciseId));
-
-    List<SampleUnit> sampleUnits = sampleService.findSampleUnitsBySampleSummary(sampleSummaryId);
-
-    if (sampleUnits.isEmpty()) {
-      LOG.info(
-          "No sample unit groups to distribute for summary",
-          kv("sampleSummaryId", sampleSummaryId));
+    LOG.info("Distributing sample units to case", kv("sampleSummaryId", sampleSummaryId));
+    try {
+      sampleSummaryDistributionService.distribute(sampleSummaryId);
+    } catch (UnknownSampleSummaryException | NoSampleUnitsInSampleSummaryException e) {
       return ResponseEntity.badRequest().build();
     }
-
-    // Catch errors distributing sample units so that only failing units are stopped
-    sampleUnits.forEach(
-        sampleUnit -> {
-          try {
-            sampleService.distributeSampleUnit(collectionExerciseId, sampleUnit);
-          } catch (RuntimeException ex) {
-            LOG.error("Failed to distribute sample unit", ex);
-          }
-        });
 
     return ResponseEntity.ok().build();
   }
