+++ src/main/java/uk/gov/ons/ctp/response/sample/service/NoSampleUnitsInSampleSummaryException.java
@@ -1,0 +1,3 @@
+package uk.gov.ons.ctp.response.sample.service;
+
+public class NoSampleUnitsInSampleSummaryException extends Exception {}
