+++ src/main/java/uk/gov/ons/ctp/response/sample/service/SampleService.java
@@ -20,13 +20,11 @@ import uk.gov.ons.ctp.response.sample.domain.model.SampleSummary;
 import uk.gov.ons.ctp.response.sample.domain.model.SampleUnit;
 import uk.gov.ons.ctp.response.sample.domain.repository.SampleSummaryRepository;
 import uk.gov.ons.ctp.response.sample.domain.repository.SampleUnitRepository;
-import uk.gov.ons.ctp.response.sample.message.SampleUnitPublisher;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO.SampleEvent;
 import uk.gov.ons.ctp.response.sample.representation.SampleSummaryDTO.SampleState;
 import uk.gov.ons.ctp.response.sample.representation.SampleUnitDTO.SampleUnitEvent;
 import uk.gov.ons.ctp.response.sample.representation.SampleUnitDTO.SampleUnitState;
-import uk.gov.ons.ctp.response.sample.representation.SampleUnitParentDTO;
 
 @Service
 @Configuration
@@ -36,8 +34,6 @@ public class SampleService {
   @Autowired private SampleSummaryRepository sampleSummaryRepository;
 
   @Autowired private SampleUnitRepository sampleUnitRepository;
-
-  @Autowired private SampleUnitPublisher sampleUnitPublisher;
 
   @Autowired
   @Qualifier("sampleSummaryTransitionManager")
@@ -313,38 +309,5 @@ public class SampleService {
     } else {
       throw new UnknownSampleSummaryException();
     }
-  }
-
-  /**
-   * Distribute a SampleUnit to get the case created in the case service. It's serialised using
-   * SampleUnitParentDTO as the SampleUnit doesn't have the collection exercise information with it.
-   *
-   * @param collectionExerciseId Collection exercise id for the sample unit
-   * @param sampleUnit for which to distribute sample units
-   */
-  public void distributeSampleUnit(String collectionExerciseId, SampleUnit sampleUnit) {
-    SampleUnitParentDTO parent = createSampleUnitParentDTOObject(collectionExerciseId, sampleUnit);
-    sampleUnitPublisher.sendSampleUnitToCase(parent);
-  }
-
-  /**
-   * Takes a populated sample unit and collection exercise ID and returns a populated
-   * sampleUnitParentDTO object
-   *
-   * @param collectionExerciseId A collection exercise ID
-   * @param sampleUnit A populated sample unit
-   * @return A populated SampleUnitParent Object
-   */
-  public SampleUnitParentDTO createSampleUnitParentDTOObject(
-      String collectionExerciseId, SampleUnit sampleUnit) {
-    SampleUnitParentDTO parent = new SampleUnitParentDTO();
-    parent.setActiveEnrolment(sampleUnit.isActiveEnrolment());
-    parent.setId(sampleUnit.getId());
-    parent.setSampleUnitRef(sampleUnit.getSampleUnitRef());
-    parent.setSampleUnitType(sampleUnit.getSampleUnitType());
-    parent.setPartyId(sampleUnit.getPartyId());
-    parent.setCollectionInstrumentId(sampleUnit.getCollectionInstrumentId());
-    parent.setCollectionExerciseId(collectionExerciseId);
-    return parent;
   }
 }
